Reorganizes code by separating municipalities and cadastral parcels in different files, restructures the tempates directory, adds a map section generator function
