+++ src/index.ts
@@ -1,6 +1,6 @@
 import { Elysia, t } from "elysia"
 import { html, Html } from "@elysiajs/html"
-import { Home } from "./templates/Home"
+import { Home } from "./templates/pages/Home"
 import { Credits } from "./namespaces/Credits"
 import { staticPlugin } from "@elysiajs/static"
 import { Phytosanitary } from "./namespaces/Phytosanitary"
@@ -12,6 +12,7 @@ import { generateDocumentation } from "./page-generators/generateDocumentation"
 import { applyRequestConfiguration } from "./applyRequestConfiguration"
 import { Seeds } from "./namespaces/Seeds"
 import { Production } from "./namespaces/Production"
+import { errorToStatus } from "./types/HTTPErrors"
 
 const PORT = import.meta.env.PORT as string
