+++ src/namespaces/Phytosanitary.ts
@@ -4,7 +4,7 @@ import { generateTablePage } from "../page-generators/generateTablePage"
 import { Hypermedia, HypermediaList } from "../Hypermedia"
 import { Field } from "../templates/components/Form"
 import { ObjectFlatMap } from "../utils"
-import { AutoList } from "../templates/AutoList"
+import { AutoList } from "../templates/views/AutoList"
 import { CreditTable } from "./Credits"
 import type { Translator } from "../Translator"
 import type { Context } from "../types/Context"
