+++ src/templates/pages/Documentation.tsx
@@ -1,5 +1,5 @@
 import { Html } from "@elysiajs/html"
-import { type HypermediaType } from "../Hypermedia"
+import { type HypermediaType } from "../../Hypermedia"
 
 interface Props {
   title: string
@@ -19,19 +19,11 @@ export function Documentation(props: Props) {
         <meta charset="utf-8" />
         <meta name="viewport" content="width=device-width, initial-scale=1" />
         <title>{props.documentation.info.title}</title>
-        <meta
-          name="description"
-          content={props.documentation.info.description}
-        />
-        <meta
-          name="og:description"
-          content={props.documentation.info.description}
-        />
+        <meta name="description" content={props.documentation.info.description} />
+        <meta name="og:description" content={props.documentation.info.description} />
       </head>
       <body>
-        <script id="api-reference">
-          {JSON.stringify(props.documentation)}
-        </script>
+        <script id="api-reference">{JSON.stringify(props.documentation)}</script>
         <script
           src={`https://cdn.jsdelivr.net/npm/@scalar/api-reference@${"latest"}/dist/browser/standalone.min.js`}
           crossorigin="true"
