+++ src/templates/views/AutoList.tsx
@@ -1,7 +1,7 @@
 import { html, Html } from "@elysiajs/html"
-import type { HypermediaType } from "../Hypermedia"
-import { Layout } from "./layouts/Layout"
-import type { Translator } from "../Translator"
+import type { HypermediaType } from "../../Hypermedia"
+import { Layout } from "../layouts/Layout"
+import type { Translator } from "../../Translator"
 
 interface PageData {
   page: {
