+++ src/templates/views/MapPage.tsx
@@ -1,8 +1,8 @@
 import { Html } from "@elysiajs/html"
-import { Layout } from "./layouts/Layout"
-import type { HypermediaType } from "../Hypermedia"
-import type { Geometry } from "../types/Geometry"
-import { Map } from "./components/Map"
+import { Layout } from "../layouts/Layout"
+import type { HypermediaType } from "../../Hypermedia"
+import type { Geometry } from "../../types/Geometry"
+import { Map } from "./../components/Map"
 
 type Coordinates = {
   latitude: number
